tools/school-synthetic-data/requirements.txt
# School Synthetic Data Generator - Dependencies

# Core dependencies
faker==22.0.0               # Generate realistic fake data
requests==2.31.0            # HTTP client for API calls
pyyaml==6.0.1              # Configuration file parsing
python-dotenv==1.0.1       # Environment variable management
click==8.1.7               # CLI framework
rich==13.7.0               # Beautiful terminal output
pydantic==2.5.3            # Data validation

# Development dependencies
pytest==7.4.3              # Testing framework
pytest-cov==4.1.0          # Code coverage
black==23.12.1             # Code formatting
flake8==7.0.0              # Linting
mypy==1.8.0                # Type checking
